+++ app/services/availabilities/public_availabilities_service.rb
@@ -1,8 +1,6 @@
 # frozen_string_literal: true
 
 # Provides helper methods for public calendar of Availability
-# FIXME, Availabilities::StatusService was refactored
-# TODO, Use Availabilities::AvailabilitiesService
 class Availabilities::PublicAvailabilitiesService
   def initialize(current_user)
     @current_user = current_user
@@ -10,95 +8,39 @@ class Availabilities::PublicAvailabilitiesService
   end
 
   # provides a list of slots and availabilities for the machines, between the given dates
-  def machines(start_date, end_date, reservations, machine_ids)
-    availabilities = Availability.includes(:tags, :machines)
-                                 .where(available_type: 'machines')
-                                 .where('start_at >= ? AND end_at <= ?', start_date, end_date)
-                                 .where(lock: false)
+  def machines(window, machine_ids, level)
+    machine_ids = [] if machine_ids.nil?
+    service = Availabilities::AvailabilitiesService.new(@current_user, level)
     slots = []
-    availabilities.each do |a|
-      slot_duration = a.slot_duration || Setting.get('slot_duration').to_i
-      a.machines.each do |machine|
-        next unless machine_ids&.include?(machine.id.to_s)
-
-        ((a.end_at - a.start_at) / slot_duration.minutes).to_i.times do |i|
-          slot = Slot.new(
-            start_at: a.start_at + (i * slot_duration).minutes,
-            end_at: a.start_at + (i * slot_duration).minutes + slot_duration.minutes,
-            availability_id: a.id,
-            availability: a,
-            machine: machine,
-            title: machine.name
-          )
-          slot = @service.machine_reserved_status(slot, reservations, @current_user)
-          slots << slot
-        end
-      end
+    machine_ids.each do |machine_id|
+      machine = Machine.friendly.find(machine_id)
+      slots.concat(service.machines(machine, @current_user, window))
     end
-    { availabilities: availabilities, slots: slots }
+    slots
   end
 
   # provides a list of slots and availabilities for the spaces, between the given dates
-  def spaces(start_date, end_date, reservations, available_id)
-    availabilities = Availability.includes(:tags, :spaces).where(available_type: 'space')
-                                 .where('start_at >= ? AND end_at <= ?', start_date, end_date)
-                                 .where(lock: false)
-
-    availabilities.where(available_id: available_id) if available_id
-
+  def spaces(window, spaces_ids, level)
+    spaces_ids = [] if spaces_ids.nil?
+    service = Availabilities::AvailabilitiesService.new(@current_user, level)
     slots = []
-    availabilities.each do |a|
-      slot_duration = a.slot_duration || Setting.get('slot_duration').to_i
-      space = a.spaces.first
-      ((a.end_at - a.start_at) / slot_duration.minutes).to_i.times do |i|
-        next unless (a.start_at + (i * slot_duration).minutes) > DateTime.current
-
-        slot = Slot.new(
-          start_at: a.start_at + (i * slot_duration).minutes,
-          end_at: a.start_at + (i * slot_duration).minutes + slot_duration.minutes,
-          availability_id: a.id,
-          availability: a,
-          space: space,
-          title: space.name
-        )
-        slot = @service.space_reserved_status(slot, reservations, @current_user)
-        slots << slot
-      end
+    spaces_ids.each do |space_id|
+      space = Space.friendly.find(space_id)
+      slots.concat(service.spaces(space, @current_user, window))
     end
-    { availabilities: availabilities, slots: slots }
+    slots
   end
 
-  def public_availabilities(start_date, end_date, reservations, ids)
-    if in_same_day(start_date, end_date)
-      # request for 1 single day
+  def public_availabilities(window, ids, events = false)
+    level = in_same_day(window[:start], window[:end]) ? 'slot' : 'availability'
+    service = Availabilities::AvailabilitiesService.new(@current_user, level)
 
-      # trainings, events
-      training_event_availabilities = Availability.includes(:tags, :trainings, :slots)
-                                                  .where(available_type: %w[training event])
-                                                  .where('start_at >= ? AND end_at <= ?', start_date, end_date)
-                                                  .where(lock: false)
-      # machines
-      machines_avail = machines(start_date, end_date, reservations, ids[:machines])
-      machine_slots = machines_avail[:slots]
-      # spaces
-      spaces_avail = spaces(start_date, end_date, reservations, ids[:spaces])
-      space_slots = spaces_avail[:slots]
+    machines_slots = machines(window, ids[:machines], level)
+    spaces_slots = spaces(window, ids[:spaces], level)
+    trainings_slots = service.trainings(Training.where(id: ids[:trainings]), @current_user, window)
+    events_slots = events ? service.events(Event.all, @current_user, window) : []
 
-      [].concat(training_event_availabilities).concat(machine_slots).concat(space_slots)
-    else
-      # request for many days (week or month)
-      avails = Availability.includes(:tags, :machines, :trainings, :spaces, :event, :slots)
-                           .where('start_at >= ? AND end_at <= ?', start_date, end_date)
-                           .where(lock: false)
-      avails.each do |a|
-        if a.available_type == 'training' || a.available_type == 'event'
-          a = @service.training_event_reserved_status(a, reservations, @current_user)
-        elsif a.available_type == 'space'
-          a.is_reserved = @service.reserved_availability?(a, @current_user)
-        end
-      end
-      avails
-    end
+    [].concat(trainings_slots).concat(events_slots).concat(machines_slots).concat(spaces_slots)
   end
 
   private