+++ test/services/subscription_extension_after_reservation_test.rb
@@ -14,15 +14,18 @@ class SubscriptionExtensionAfterReservationTest < ActiveSupport::TestCase
 
     @user.reservations.destroy_all # ensure no reservations
 
+    @slot_reservation_machine = SlotsReservation.new({ slot_id: @machine.availabilities.first.slots.first.id })
+    @slot_reservation_training = SlotsReservation.new({ slot_id: @training.availabilities.first.slots.first.id })
+
     @reservation_machine = Reservation.new(
       statistic_profile: @user.statistic_profile,
       reservable: @machine,
-      slots_reservations: [{ slot_id: @machine.availabilities.first.slots.first.id }]
+      slots_reservations: [@slot_reservation_machine]
     )
     @reservation_training = Reservation.new(
       statistic_profile: @user.statistic_profile,
       reservable: @training,
-      slots_reservations: [{ slot_id: @training.availabilities.first.slots.first.id }]
+      slots_reservations: [@slot_reservation_training]
     )
     @reservation_training.save!
   end
@@ -53,6 +56,6 @@ class SubscriptionExtensionAfterReservationTest < ActiveSupport::TestCase
 
   test 'method extend_subscription' do
     SubscriptionExtensionAfterReservation.new(@reservation_training).extend_subscription
-    assert_equal @reservation_training.slots.first.start_at + @plan.duration, @user.subscription.expired_at
+    assert_equal @reservation_training.slots_reservations.first.slot.start_at + @plan.duration, @user.subscription.expired_at
   end
 end